Move planner components around
