+++ src/components/planner/PlannerCoursesTable/PlannerCard.tsx
